// types/contracts/MockTerminus.ts
export interface MockTerminus extends BaseContract {
  constructor(
    jsonInterface: any[],
    address?: string,
    options?: ContractOptions
  ): MockTerminus;
  clone(): MockTerminus;
  methods: {
    approveForPool(
      poolID: number | string | BN,
      operator: string
    ): NonPayableTransactionObject<void>;

    balanceOf(
      account: string,
      id: number | string | BN
    ): NonPayableTransactionObject<string>;

    balanceOfBatch(
      accounts: string[],
      ids: (number | string | BN)[]
    ): NonPayableTransactionObject<string[]>;

    burn(
      from: string,
      poolID: number | string | BN,
      amount: number | string | BN
    ): NonPayableTransactionObject<void>;

    contractURI(): NonPayableTransactionObject<string>;

    createPoolV1(
      _capacity: number | string | BN,
      _transferable: boolean,
      _burnable: boolean
    ): NonPayableTransactionObject<string>;

    createSimplePool(
      _capacity: number | string | BN
    ): NonPayableTransactionObject<string>;

    isApprovedForAll(
      account: string,
      operator: string
    ): NonPayableTransactionObject<boolean>;

    isApprovedForPool(
      poolID: number | string | BN,
      operator: string
    ): NonPayableTransactionObject<boolean>;

    mint(
      to: string,
      poolID: number | string | BN,
      amount: number | string | BN,
      data: string | number[]
    ): NonPayableTransactionObject<void>;

    mintBatch(
      to: string,
      poolIDs: (number | string | BN)[],
      amounts: (number | string | BN)[],
      data: string | number[]
    ): NonPayableTransactionObject<void>;

    paymentToken(): NonPayableTransactionObject<string>;

    poolBasePrice(): NonPayableTransactionObject<string>;

    poolMintBatch(
      id: number | string | BN,
      toAddresses: string[],
      amounts: (number | string | BN)[]
    ): NonPayableTransactionObject<void>;

    safeBatchTransferFrom(
      from: string,
      to: string,
      ids: (number | string | BN)[],
      amounts: (number | string | BN)[],
      data: string | number[]
    ): NonPayableTransactionObject<void>;

    safeTransferFrom(
      from: string,
      to: string,
      id: number | string | BN,
      amount: number | string | BN,
      data: string | number[]
    ): NonPayableTransactionObject<void>;

    setApprovalForAll(
      operator: string,
      approved: boolean
    ): NonPayableTransactionObject<void>;

    setContractURI(_contractURI: string): NonPayableTransactionObject<void>;

    setController(newController: string): NonPayableTransactionObject<void>;

    setPaymentToken(newPaymentToken: string): NonPayableTransactionObject<void>;

    setPoolBasePrice(
      newBasePrice: number | string | BN
    ): NonPayableTransactionObject<void>;

    setPoolController(
      poolID: number | string | BN,
      newController: string
    ): NonPayableTransactionObject<void>;

    setURI(
      poolID: number | string | BN,
      poolURI: string
    ): NonPayableTransactionObject<void>;

    supportsInterface(
      interfaceId: string | number[]
    ): NonPayableTransactionObject<boolean>;

    terminusController(): NonPayableTransactionObject<string>;

    terminusPoolCapacity(
      poolID: number | string | BN
    ): NonPayableTransactionObject<string>;

    terminusPoolController(
      poolID: number | string | BN
    ): NonPayableTransactionObject<string>;

    terminusPoolSupply(
      poolID: number | string | BN
    ): NonPayableTransactionObject<string>;

    totalPools(): NonPayableTransactionObject<string>;

    uri(poolID: number | string | BN): NonPayableTransactionObject<string>;

    withdrawPayments(
      toAddress: string,
      amount: number | string | BN
    ): NonPayableTransactionObject<void>;
  };
// ...
}